% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.sim.three.R
\name{anova.sim.three}
\alias{anova.sim.three}
\title{anova.sim.three}
\usage{
\method{anova}{sim.three}(group_n, f_sqr, sd.1)
}
\arguments{
\item{group_n}{a (non-empty) numeric value of desired sample size per group}

\item{f_sqr}{a (non-empty) numeric value of desired cohen's f squared value}

\item{sd.1}{a (non-empty) numeric value of desired standard deviation ratio}
}
\value{
a dataframe
}
\description{
simulate three-group anova data
}
