% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DACF-package.R
\docType{package}
\name{DACF-package}
\alias{DACF-package}
\alias{DACF}
\title{Data Analyses for data with Ceiling/Floor effects}
\description{
The \code{DACF} package is used to conduct data analyses for data with
ceiling/floor effects. The current version contains functions to
estimate mean and variances and to conduct mean comparison tests.
}
\section{Usage}{

To access this package's tutorial, type the following line into the console:

\code{vignette("DACF-vignette")}
}

