% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F.star.test.R
\name{f.star.test}
\alias{f.star.test}
\title{f.star.test}
\usage{
f.star.test(means, variances, ns)
}
\arguments{
\item{means}{a (non-empty) numeric vector of the group means}

\item{variances}{a (non-empty) numeric vector of the group variances}

\item{ns}{a (non-empty) numeric vector of sample size per group}
}
\value{
F star value, p value, and f.squared
}
\description{
conduct a Browne-Forsythe F star test
}
