% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce.cfe.R
\name{induce.cfe}
\alias{induce.cfe}
\title{induce.cfe}
\usage{
induce.cfe(floor.perc, ceiling.perc, y)
}
\arguments{
\item{floor.perc}{a (non-empty) numeric value from 0 to 1 denoting the desired percentage of floor effects}

\item{ceiling.perc}{a (non-empty) numeric value from 0 to 1 denoting the desired percentage of ceiling effects}

\item{y}{a (non-empty) numeric vector of data}
}
\value{
y scores with induced ceiling/floor effects
}
\description{
inducing ceiling/floor effects in data
}
