% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lw.t.test.R
\name{lw.t.test}
\alias{lw.t.test}
\title{lw.t.test}
\usage{
lw.t.test(x1, x2, method_type)
}
\arguments{
\item{x1}{a (non-empty) numeric vector of data values for group 1 with floor/ceiling effects}

\item{x2}{a (non-empty) numeric vector of data values for group 2 with floor/ceiling effects}

\item{method_type}{a character string specifying the preferred method type. "a" uses the original sample size and "b" uses after-truncation sample size.}
}
\value{
the test statistic, p value, effect size estimate, and confidence interval for t test adjusting for ceiling/floor effects
}
\description{
conduct a t test adjusting for ceiling and/or floor effects
}
