% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec.mean.var.R
\name{rec.mean.var}
\alias{rec.mean.var}
\title{rec.mean.var}
\usage{
rec.mean.var(y)
}
\arguments{
\item{y}{a (non-empty) numeric vector of data with ceiling/floor effects}
}
\value{
percentage of ceiling/floor and recovered mean and variance
}
\description{
recover mean and variance of the data with ceiling/floor effects
}
