% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeganova.sim.R
\name{threeganova.sim}
\alias{threeganova.sim}
\title{threeganova.sim}
\usage{
threeganova.sim(group_n, f_sqr, sd.1)
}
\arguments{
\item{group_n}{a (non-empty) numeric value of desired sample size per group}

\item{f_sqr}{a (non-empty) numeric value of desired Cohen's f squared value}

\item{sd.1}{a (non-empty) numeric value of desired standard deviation ratio}
}
\value{
a dataframe containing scores "y", grouping factor "group", and residual errors.
}
\description{
simulate three-group anova data
}
